Doc: Add doc for UncaughtException
